pyproject.toml
[tool.poetry]
name = "aqua-tally-backend"
version = "0.1.0"
description = ""
authors = ["do-ald533 <douglas.aldred533@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "aqua_tally_backend"}]

[tool.poe.tasks.django]
cmd = "python manage.py"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.3"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
poethepoet = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
